+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## [Unreleased]
 ### Added
 - can now hide commands in `/help`
+- added logging console
 ### Fixed
 - preventing the bot from deleting every messages starting with the prefix (the one in [2.1-a2] does not work)