+++ CHANGELOG.md
@@ -3,10 +3,35 @@
 
 ## [Unreleased]
 ### Added
-- ability to unload, load or reload cogs during runtime
+- can now hide commands in `/help`
+- logging console
+- debug / fix mode
+- ability to unload, reload and load commands
+- new warning system that replace the old one and have less security issues
+- command error handler
+### Fixed
+- preventing the bot from deleting every messages starting with the prefix (the one in [2.1-a2] does not work)
 
+## [2.1-a2] - 2020-07-05
+### Added
+- allowing users to end all polls in a channel with `/poll end all` (buggy)
+- eval command for bot owner only
+- compressed spamming warning system, boosted performance
+- compressed bot.py
+- bot owner can now run `/eval` python expressions.
+### Changed
+- compressed the code for a bit without losing readability and performance
+- QuickPoll now have special return code representing different errors
+- the python shell will now print out the whole traceback
+### Removed
+- ability for not specifying the amount of messages to read for `/poll check`
+- the `/table` command
+### Fixed
+- check polls might compare str with int
+- ~~preventing the bot from deleting every messages starting with `/` (which is the prefix)~~
+- bot might read its own error messages in the python debugging channel
 
-## [2.1-a1]
+## [2.1-a1] - 2020-07-02
 ### Added
 - allowing users to add custom shortcuts for `/mc srv`
 - allowing moderators to add `"`s in warnings reasons
@@ -15,7 +40,7 @@
 ### Removed
 - sudo mode (moderating)
 
-## [2.1-a]
+## [2.1-a] - 2020-06-22
 ### Added
 - Karma system integrated with [Merlin-js]
 - more fun commands
@@ -30,9 +55,7 @@
 - no more verbose mode :O
 
 
-[Unreleased]:   https://github.com/windowsboy111/Merlin-py/compare/3.0-a...HEAD
-[3.0-a1]:       https://github.com/windowsboy111/Merlin-py/compare/2.0.0...3.0-a
-[2.1-a]:        https://github.com/windowsboy111/Merlin-py/compare/2.0.0...3.0-a
-[2.0.0]:        https://github.com/windowsboy111/Merlin-py/releases/tag/2.0.0
-
-[Merlin-js]:    https://github.com/windowsboy111/Merlin-js
+[Unreleased]:   https://github.com/windowsboy111/Merlin-py/compare/2.1-a2...HEAD
+[2.1-a2]:       https://github.com/windowsboy111/Merlin-py/compare/2.1-a1...2.1-a2
+[2.1-a1]:       https://github.com/windowsboy111/Merlin-py/compare/2.1a...2.1-a1
+[2.1-a]:        https://github.com/windowsboy111/Merlin-py/compare/2.0.0...2.1a